Separation in modules refactors, tidy code
Changes:
- Separated classes in modules.
- Removed innecessary comments.
