--- data_models/TradeAssistant.py
@@ -1,11 +1,0 @@
-import json
-from typing import List
-# import recognizers_suite as Recognizers
-# from recognizers_suite import Culture, ModelResult
-
-from datetime import datetime
-
-from collections import deque
-
-import random
-
